Modify client side code to use common serialization library
